+++ trixnity-messenger/src/commonMain/kotlin/de/connect2x/trixnity/messenger/viewmodel/room/settings/RoomSettingsViewModel.kt
@@ -48,6 +48,7 @@ interface RoomSettingsViewModel {
     val roomSettingsNotificationsViewModel: RoomSettingsNotificationsViewModel
     val roomSettingsHistoryVisibilityViewModel: RoomSettingsHistoryVisibilityViewModel
     val roomSettingsJoinRulesViewModel : RoomSettingsJoinRulesViewModel
+    val roomSettingsSecurityViewModel: RoomSettingsSecurityViewModel
     val roomSettingsAliasViewModel: RoomSettingsAliasViewModel
     val leaveRoomSettingEntryText: StateFlow<String>
     val leaveRoomWarningOpen: StateFlow<Boolean>
@@ -106,6 +107,11 @@ class RoomSettingsViewModelImpl(
 
     override val roomSettingsJoinRulesViewModel: RoomSettingsJoinRulesViewModel by lazy {
         get<RoomSettingsJoinRulesViewModelFactory>()
+            .create(viewModelContext, selectedRoomId, error)
+    }
+
+    override val roomSettingsSecurityViewModel: RoomSettingsSecurityViewModel by lazy {
+        get<RoomSettingsSecurityViewModelFactory>()
             .create(viewModelContext, selectedRoomId, error)
     }
 
@@ -208,6 +214,8 @@ class PreviewRoomSettingsViewModel : RoomSettingsViewModel {
         = PreviewRoomSettingsHistoryVisibilityViewModel()
     override val roomSettingsJoinRulesViewModel: PreviewRoomSettingsJoinRulesViewModel
         = PreviewRoomSettingsJoinRulesViewModel()
+    override val roomSettingsSecurityViewModel: PreviewRoomSettingsSecurityViewModel
+        = PreviewRoomSettingsSecurityViewModel()
     override val roomSettingsAliasViewModel: RoomSettingsAliasViewModel
         = PreviewRoomSettingsAliasViewModel()
     override val error: MutableStateFlow<String?> = MutableStateFlow(null)