+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Setting to enable E2E encryption in rooms
+
 ### Changed
 
 - upgraded to Kotlin 2.0
@@ -18,7 +20,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - improvements in ExportRoom: parallel decryption, collect decryption errors
 - move gradle locks to CI
 - restrict ability to change room history to WORLD_READABLE when the group is encrypted
-
+- fixed typo in RoomEncryptionEnabledViewModel
 
 ### Deprecated